mkdtimg: Bundle mkdtimg from AOSP's libufdt
The python rewrite of mkdtimg, mtdtboimg.py, isn't feature-complete, so it
can't replace the C version of mkdtimg. Since the right version of mkdtimg
is not something that everyone just happens to have in their environment,
bundle the C version of mkdtimg from AOSP's libufdt and use that instead of
relying on mkdtimg from the environment.

This depends upon libfdt from aosp-dtc.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>